abi-gen/TS: fix jsdoc alignment
Before this change, docs were being rendered like:

```
    * some jsdoc text blah blah.
     * @param x some param
    * @returns whatever
    */
```

No corresponding change was needed to the test-cli/output because
prettier is run on those files after abi-gen produces them, and prettier
was cleaning up this alignment for us.  I hit this as a tslint error
when using abi-gen elsewhere.  (I wasn't using prettier on the generated
code in that usage.)