contract-artifacts: fix bug transforming non-json
I was trying to run `yarn transform_artifacts` and was getting an
exception that didn't give any indication of the problem.

After adding some debug trace, I found out that the culprit was a
`.DevUtils.json.swp` file in the artifacts directory, a byproduct of my
having the file open in vim.

This change ensures that only files that END with `.json` will be
transformed.